# feat: improve JavaScript performance by caching the runtime + polyfill