--- a/kits/javascript/Cargo.toml
+++ b/kits/javascript/Cargo.toml
@@ -11,3 +11,4 @@ javy = { version = "2.0.0", features = ["json"] }
 regex = "1.10.1"
 # Use an old version until we add support for components.
 wit-bindgen-rust = { git = "https://github.com/bytecodealliance/wit-bindgen", rev = "cb871cfa1ee460b51eb1d144b175b9aab9c50aba" }
+once_cell = "1.18.0"