--- a/kits/javascript/Makefile
+++ b/kits/javascript/Makefile
@@ -1,5 +1,7 @@
-.PHONY: build
+.PHONY: build build-src
 
-build:
-	cargo build --target wasm32-wasi --release && \
-	cp ./target/wasm32-wasi/release/wasm-workers-quick-js-engine.wasm ./
+build-src:
+	cargo build --target wasm32-wasi --release
+
+build: build-src
+	wizer --allow-wasi --wasm-bulk-memory true -o ./wasm-workers-quick-js-engine.wasm ./target/wasm32-wasi/release/wasm-workers-quick-js-engine.wasm