--- a/kits/javascript/src/error.rs
+++ b/kits/javascript/src/error.rs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 /// List of runtime errors
+#[derive(Debug)]
 pub enum RuntimeError {
     InvalidBinding { invalid_export: String },
 }